isOrdered to QR
